refactor: remove dead src/tools/spf-analysis.ts fork
src/tools/spf-analysis.ts was a full duplicate of the live
packages/dns-checks/src/checks/spf-analysis.ts, imported by nothing but its
own test/spf-analysis.spec.ts — and it never got the v3.15.1 RFC 7208 §4.6.4
redirect= lookup-count fix, so the fork's lookup budget silently diverged.

- Delete src/tools/spf-analysis.ts (no Worker source imports it; verified).
- Delete test/spf-analysis.spec.ts (only exercised the dead fork; its
  analyzeSpfLookupBudget case asserted the old undercount).
- Relocate its still-unique coverage (extractSpfSignalDomains/extractLookupDomains
  and checkBroadIpRanges — previously tested ONLY against the fork) onto the live
  copy in packages/dns-checks/src/__tests__/checks/check-spf.test.ts.

No runtime change (the fork was dead). typecheck + lint clean; dns-checks 389,
full Worker suite green.
